+++ caching-service/src/main/java/org/zowe/apiml/caching/service/vsam/ZFileProducer.java
@@ -12,8 +12,8 @@ package org.zowe.apiml.caching.service.vsam;
 import lombok.RequiredArgsConstructor;
 import org.zowe.apiml.caching.service.vsam.config.VsamConfig;
 import org.zowe.apiml.message.log.ApimlLogger;
-import org.zowe.apiml.util.ClassOrDefaultProxyUtils;
-import org.zowe.apiml.zfile.*;
+// import org.zowe.apiml.util.ClassOrDefaultProxyUtils;
+import com.ibm.jzos.*;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.regex.Pattern;
@@ -27,33 +27,12 @@ public class ZFileProducer {
     private final ApimlLogger apimlLog;
 
     @SuppressWarnings({"java:S1130"}) // It's necessary for the zOS implementations.
-    public ZFile openZfile() throws VsamRecordException {
+    public ZFile openZfile() throws VsamRecordException, ClassNotFoundException, InstantiationException, IllegalAccessException, NoSuchMethodException, InvocationTargetException {
         if (!REGEX_CORRECT_FILENAME.matcher(vsamConfig.getFileName()).find()) {
             throw new IllegalStateException("VsamFile does not exist");
         }
         String calledMethod = "getMessage";
-        try {
-            return ClassOrDefaultProxyUtils.createProxyByConstructor(ZFile.class, "com.ibm.jzos.ZFile",
-                ZFileDummyImpl::new,
-                new Class[]{String.class, String.class, int.class},
-                new Object[]{vsamConfig.getFileName(), options.getOptionsString(), ZFileConstants.FLAG_DISP_SHR + ZFileConstants.FLAG_PDS_ENQ},
-                new ClassOrDefaultProxyUtils.ByMethodName<>(
-                    "com.ibm.jzos.ZFileException", ZFileException.class,
-                    "getFileName", calledMethod, "getErrnoMsg", "getErrno", "getErrno2", "getLastOp", "getAmrcBytes",
-                    "getAbendCode", "getAbendRc", "getFeedbackRc", "getFeedbackFtncd", "getFeedbackFdbk"),
-                new ClassOrDefaultProxyUtils.ByMethodName<>(
-                    "com.ibm.jzos.RcException", RcException.class,
-                    calledMethod, "getRc"),
-                new ClassOrDefaultProxyUtils.ByMethodName<>(
-                    "com.ibm.jzos.EnqueueException", EnqueueException.class,
-                    calledMethod, "getRc")
-            );
-        } catch (ClassNotFoundException | InstantiationException | IllegalAccessException | NoSuchMethodException e) {
-            apimlLog.log("org.zowe.apiml.cache.noJzosImplementation");
-
-            throw new JzosImplementationException(e);
-        } catch (InvocationTargetException e) {
-            throw new VsamRecordException("Failed opening of file", e.getTargetException());
-        }
+        ZFile zfile = new ZFile();
+        return zfile;
     }
 }
