+++ settings.gradle
@@ -51,4 +51,5 @@ include 'apiml-tomcat-common'
 include 'onboarding-enabler-java-sample-app-plain-java'
 include 'apiml-sample-extension'
 include 'apiml-sample-extension-package'
-
+include 'mock-zos'
+//include ':progressfragment'
